auth_qr/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
cryptography
qrcode
pillow
python-jose
passlib[bcrypt]
pydantic[email]
